Migrate upload function to GeoGuessr API v4
Uploading broke due to the introduction of strict versioning of maps by
GeoGuessr. This restores the upload functionality.

Also adds support for the new map tags feature introduced in a recent v4
revision, replacing the UI for the useless `regions` property (which is
only used for polygonal maps, and not handpicked maps).

Importing continues to use a v3 endpoint, as it still works, and the v4
drafts endpoint seems to require some kind of a draft initalization to
occur before returning correct map data.

There is a regression: Toggling off the publish / visibility to others
checkbox and then uploading fails to unpublish the map, as the v4
endpoint ignores the `published` property.